+++ cli/src/commands/git/push.rs
@@ -30,6 +30,7 @@ use jj_lib::git;
 use jj_lib::git::GitBranchPushTargets;
 use jj_lib::object_id::ObjectId;
 use jj_lib::op_store::RefTarget;
+use jj_lib::op_store::RemoteRef;
 use jj_lib::refs::classify_bookmark_push_action;
 use jj_lib::refs::BookmarkPushAction;
 use jj_lib::refs::BookmarkPushUpdate;
@@ -48,7 +49,9 @@ use crate::cli_util::CommandHelper;
 use crate::cli_util::RevisionArg;
 use crate::cli_util::WorkspaceCommandHelper;
 use crate::cli_util::WorkspaceCommandTransaction;
+use crate::command_error::cli_error;
 use crate::command_error::user_error;
+use crate::command_error::user_error_with_hint;
 use crate::command_error::CommandError;
 use crate::commands::git::get_single_remote;
 use crate::complete;
@@ -79,7 +82,7 @@ use crate::ui::Ui;
 ///     https://jj-vcs.github.io/jj/latest/bookmarks/#conflicts
 
 #[derive(clap::Args, Clone, Debug)]
-#[command(group(ArgGroup::new("specific").args(&["bookmark", "change", "revisions"]).multiple(true)))]
+#[command(group(ArgGroup::new("specific").args(&["bookmark", "change", "revisions", "named"]).multiple(true)))]
 #[command(group(ArgGroup::new("what").args(&["all", "deleted", "tracked"]).conflicts_with("specific")))]
 pub struct GitPushArgs {
     /// The remote to push to (only named remotes are supported)
@@ -168,6 +171,13 @@ pub struct GitPushArgs {
         add = ArgValueCandidates::new(complete::mutable_revisions)
     )]
     change: Vec<RevisionArg>,
+    /// Specify a new bookmark name and a revision to push under that name, e.g.
+    /// '--named myfeature=@'.
+    ///
+    /// Does not require --allow-new
+    // TODO: Add arg completer
+    #[arg(long, value_name = "NAME=REVISION")]
+    named: Vec<String>,
     /// Only display what will change on the remote
     #[arg(long)]
     dry_run: bool,
@@ -259,7 +269,6 @@ pub fn cmd_git_push(
             };
             (bookmark_name.as_ref(), targets)
         });
-        let view = tx.repo().view();
         for (bookmark_name, targets) in change_bookmarks {
             if !seen_bookmarks.insert(bookmark_name) {
                 continue;
@@ -275,6 +284,38 @@ pub fn cmd_git_push(
             }
         }
 
+        for name_revision in &args.named {
+            let (name, target) =
+                create_explicitly_named_bookmarks(ui, &remote, &mut tx, name_revision)?;
+
+            let allow_new = true;
+            assert_eq!(
+                tx.repo().view().get_remote_bookmark(RemoteRefSymbol {
+                    name: &name,
+                    remote: &remote,
+                }),
+                RemoteRef::absent_ref(),
+                "`create_explicitly_named_bookmarks` ensures {name}@{remote} is absent"
+            );
+            match classify_bookmark_update(
+                &name,
+                &remote,
+                LocalAndRemoteRef {
+                    local_target: &target,
+                    remote_ref: &RemoteRef::absent(),
+                },
+                allow_new,
+            ) {
+                Ok(Some(update)) => bookmark_updates.push((name, update)),
+                Ok(None) => writeln!(
+                    ui.status(),
+                    "Bookmark {name}@{remote} already matches the desired revision",
+                )?,
+                Err(reason) => return Err(reason.into()),
+            }
+        }
+
+        let view = tx.repo().view();
         let allow_new = args.allow_new || tx.settings().get("git.push-new-bookmarks")?;
         let bookmarks_by_name = find_bookmarks_to_push(view, &args.bookmark, &remote)?;
         for &(bookmark_name, targets) in &bookmarks_by_name {
@@ -291,8 +332,10 @@ pub fn cmd_git_push(
             }
         }
 
-        let use_default_revset =
-            args.bookmark.is_empty() && args.change.is_empty() && args.revisions.is_empty();
+        let use_default_revset = args.bookmark.is_empty()
+            && args.change.is_empty()
+            && args.revisions.is_empty()
+            && args.named.is_empty();
         let bookmarks_targeted = find_bookmarks_targeted_by_revisions(
             ui,
             tx.base_workspace_helper(),
@@ -674,6 +717,68 @@ fn classify_bookmark_update(
         }
         BookmarkPushAction::Update(update) => Ok(Some(update)),
     }
+}
+
+// Creates the bookmars for `--named` arguments
+//
+// The logic is not identical to that of `jj bookmark create` since we need to
+// make sure the new bookmark is safe to push.
+fn create_explicitly_named_bookmarks(
+    ui: &mut Ui,
+    remote: &String,
+    tx: &mut WorkspaceCommandTransaction<'_>,
+    name_revision: &String,
+) -> Result<(String, RefTarget), CommandError> {
+    let Some((name, revision_str)) = name_revision.split_once('=') else {
+        return Err(cli_error(format!(
+            "Argument '{name_revision}' must include '=' and have the form NAME=REVISION"
+        ))
+        .hinted("For example, `--named myfeature=@` is valid syntax"));
+    };
+    // Revisions are lenient towards spaces, so let's make the name lenient as well
+    // Git does not allow bookmark names that include spaces anyway.
+    let name = name.trim();
+    if name.is_empty() || revision_str.is_empty() {
+        return Err(cli_error(format!(
+            "Argument '{name_revision}' must have the form NAME=REVISION, with both NAME and \
+             REVISION non-empty."
+        ))
+        .hinted("For example, `--named myfeature=@` is valid syntax"));
+    }
+    if tx.repo().view().get_local_bookmark(name).is_present() {
+        return Err(user_error_with_hint(
+            format!("Bookmark already exists: '{name}'"),
+            format!(
+                "Use 'jj bookmark move' to move it, and 'jj git push -b {name} [--allow-new]' to \
+                 push it."
+            ),
+        ));
+    }
+    if tx
+        .repo()
+        .view()
+        .get_remote_bookmark(RemoteRefSymbol { name, remote })
+        .is_present()
+    {
+        // This is slightly different from classify_bookmark_update since we
+        // error even in the case where the remote branch is not tracked. This can
+        // happen if the local branch is deleted by the (untracked) remote-tracking
+        // bookmark still exists.
+        //
+        // This may not be strictly necessary, but ensuring that the bookmark is always
+        // created on the remote from nothing simplifies both the UI and the logic.
+        return Err(user_error_with_hint(
+            format!("Remote bookmark already exists: '{name}@{remote}'"),
+            "Use 'jj bookmark track' to track it.",
+        ));
+    }
+    let revision = tx
+        .base_workspace_helper()
+        .resolve_single_rev(ui, &revision_str.to_string().into())?;
+    let ref_target = RefTarget::normal(revision.id().clone());
+    tx.repo_mut()
+        .set_local_bookmark_target(name, ref_target.clone());
+    Ok((name.to_string(), ref_target))
 }
 
 /// Creates or moves bookmarks based on the change IDs.